images styling update

+++ client/src/components/Images/Images.tsx
@@ -3,6 +3,7 @@ import "./Images.css";
 import downloadIcon from "../../assets/tabler_download.svg";
 import EditIcon from "../../assets/tabler_edit.svg";
 import DeleteIcon from "../../assets/tabler_trash-x.svg";
+import { formatDate } from "../../utils/helpers";
 
 interface ImageData {
   id: string;
@@ -16,6 +17,8 @@ interface ItemsToDisplayProps {
 
 function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
   const [images, setImages] = useState<any>([]);
+  const [imageDate, setImageDate] = useState<any>([]);
+  const [loadingImages, setLoadingImages] = useState(true);
 
   const fetchImages = async () => {
     try {
@@ -25,9 +28,13 @@ function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
       }
 
       const data = await response.json();
+      console.log(data[0].date);
+      setImageDate(data[0].date);
       setImages(data);
     } catch (error) {
       console.error("Error fetching images:", error);
+    } finally {
+      setLoadingImages(false);
     }
   };
 
@@ -51,6 +58,7 @@ function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
   };
 
   const handleDeleteImage = async (id: string) => {
+    if (!window.confirm("Are you sure you want to delete this image?")) return;
     try {
       const response = await fetch(`http://localhost:4000/api/images/${id}`, {
         method: "DELETE",
@@ -71,10 +79,14 @@ function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
 
   return (
     <div>
-      <h1>Images</h1>
+      {!loadingImages && <h2 className="images-date">{formatDate(imageDate)}</h2>}
+
       <div className="images-list">
         {images.map((imageData: ImageData, index: number) => (
-          <div key={imageData.id} className="image-container">
+          <div
+            key={imageData.id}
+            className={`image-container ${loadingImages ? "image-loading" : ""}`}
+          >
             {imageData.image && (
               <section className="images">
                 <img
@@ -83,6 +95,7 @@ function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
                   alt={imageData.label}
                   width={200}
                   height={200}
+                  onLoad={() => setLoadingImages(false)}
                 />
 
                 {imageData.label && <p className="label">{imageData.label}</p>}
@@ -102,12 +115,8 @@ function ImagesDisplay({ updateData }: ItemsToDisplayProps) {
                 </section> */}
 
                 <section className="actions">
-                  <button className="action-button">
-                    <img
-                      src={DeleteIcon}
-                      alt="delete"
-                      onClick={() => handleDeleteImage(imageData.id)}
-                    />
+                  <button className="action-button" onClick={() => handleDeleteImage(imageData.id)}>
+                    <img src={DeleteIcon} alt="delete" />
                     <span>Delete</span>
                   </button>
                 </section>
